massive generation refactor
