+++ glsl.cc
@@ -31,6 +31,10 @@ const float TERRAIN_BASE_HEIGHT = float(WATER_SURROUNDING_HEIGHT + WATER_BASE_HE
 const float TERRAIN_OCEAN_THRESHOLD = OCEAN_THRESHOLD;
 const float TERRAIN_RIVER_THRESHOLD = RIVER_BASE;
 const float TERRAIN_WATER_THRESHOLD = WATER_THRESHOLD;
+// grass
+const float TERRAIN_GRASS_THRESHOLD = GRASS_THRESHOLD;
+// trees
+const float TERRAIN_TREE_THRESHOLD = VEGGIE_THRESHOLD;
 // rocks
 const float TERRAIN_ROCK_THRESHOLD = ROCK_THRESHOLD;
 // flattener
