+++ promise.h
@@ -4,7 +4,6 @@
 #include "vectorMath.h"
 #include "sync.h"
 #include "result.h"
-// #include "lock.h"
 #include <vector>
 #include <deque>
 #include <atomic>
