+++ task.h
@@ -3,20 +3,12 @@
 
 #include "vectorMath.h"
 #include "sync.h"
-#include "lock.h"
 #include "vector.h"
 #include <array>
 #include <vector>
 #include <deque>
-// #include <semaphore>
 #include <atomic>
 #include <emscripten.h>
-
-//
-
-// class DCInstance;
-
-//
 
 class Task {
 public:
@@ -28,21 +20,15 @@ public:
     int lod;
     int priority;
 
-    // Task(uint32_t id, std::function<void()> fn);
     Task(uint32_t id, int priority, std::function<void()> fn);
     Task(uint32_t id, const vm::vec3 &worldPosition, int lod, std::function<void()> fn);
     Task(uint32_t id, const vm::vec3 &worldPosition, int lod, int priority, std::function<void()> fn);
     ~Task();
 
-    // bool tryLock();
-    // void lock();
-    // void unlock();
     void run();
     void cancel();
-    // void ensurePop();
 
     int getPriority() const;
-    // Sphere getSphere() const;
     Box3 getBox() const;
 };
